Make generated visitor id random to prevent collisions
Right now, if you do not specify a visitor id, it will autogenerate
a visitor id based on a date, down to the minute, and then tack on
a random number between 10,000 and 99,999.

This is not great because it has a high probability of collisions.

For example, given ther 86,400 seconds in a day, if you have 2 new
visitors hitting flagship a second, there's a 67% chance that you'll
have a collision. See here for calculations:
https://www.calculator.net/probability-calculator.html?cal4pa=0.00001111111&cal4par=86400&cal4pb=0&cal4pbr=0&calctype=series&x=19&y=16#series

The fix here is to simply use uuids, which has a 1 in 2.71 quatillion chance
of collision.