refactor: Add cleanup method to Visitor class
This commit adds a new cleanup method to the Visitor class in the `Visitor.ts` file. The cleanup method is responsible for cleaning up any resources used by the Visitor instance. It calls the `cleanup` method of the `visitorDelegate` object and removes the event listener for the `EMIT_READY` event.